Add files via upload
Programming Language: JAVA
Development IDE: Android Studio

Tic-Tac-Toe is a classic two-player strategy game where Player 1 marks "X" and Player 2 marks "O" on a 3x3 grid. The goal is to form three consecutive symbols in a row, column, or diagonal (both positive and negative). If a player achieves this, they win. If the grid is completely filled without a winner, the game ends in a tie.

This game has been developed using Java in Android Studio. The first screen displays the title "Tic-Tac-Toe" along with a "Play" button to start the game. Clicking it navigates users to a page where they enter their names and submit them before proceeding to the game board.

The gameplay alternates between players, indicating whose turn it is. Player 1 places "X" while Player 2 places "O" on the board. The winner’s name is displayed at the top when the game ends. If there is no winner, the message "Match Tied" appears. Players can either select "Play Again" to reset the board and continue with the same names or choose "Home" to return to the main screen and start over.